+++ app/(site)/posts/[slug]/page.tsx
@@ -1,7 +1,8 @@
 import { Metadata } from "next";
 import Link from "next/link";
 import { notFound } from "next/navigation";
-import { allPosts } from "contentlayer/generated";
+import { PostSeries, PostWithSeries, SeriesItem } from "@/types";
+import { allPosts, Post } from "contentlayer/generated";
 import { format, parseISO } from "date-fns";
 import { Home } from "lucide-react";
 
@@ -10,6 +11,7 @@ import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from "@/
 import { Card, CardContent, CardFooter, CardHeader, CardTitle } from "@/components/ui/card";
 import { Separator } from "@/components/ui/separator";
 import { Mdx } from "@/components/mdx-components";
+import { PostSeriesBox } from "@/components/post-series-box";
 import { TableOfContents } from "@/components/table-of-contents";
 
 interface PostProps {
@@ -18,11 +20,28 @@ interface PostProps {
   };
 }
 
-async function getPostFromParams(params: PostProps["params"]) {
+async function getPostFromParams(params: PostProps["params"]): Promise<any> {
   const post = allPosts.find((post) => post.slug === params.slug);
 
   if (!post) {
     null;
+  }
+
+  if (post?.series) {
+    const seriesPosts: SeriesItem[] = allPosts
+      .filter((p) => p.series?.title === post.series?.title)
+      .sort((a, b) => Number(a.series!.order) - Number(b.series!.order))
+      .map((p) => {
+        return {
+          title: p.title,
+          slug: p.slug,
+          status: p.status,
+          isCurrent: p.slug === post.slug,
+        };
+      });
+    if (seriesPosts.length > 0) {
+      return { ...post, series: { ...post.series, posts: seriesPosts } as PostSeries };
+    }
   }
 
   return post;
@@ -130,6 +149,11 @@ export default async function PostPage({ params }: PostProps) {
             <p className="mb-2 mt-0 text-xl text-slate-700 dark:text-slate-200">{post.description}</p>
           )}
           <hr className="my-4" />
+          {post?.series && (
+            <div className="not-prose">
+              <PostSeriesBox data={post.series} />
+            </div>
+          )}
           <Mdx code={post.body.code} />
           <hr className="my-4" />
           {post.tags && (