+++ types/index.tsx
@@ -1,4 +1,5 @@
 import { LinkProps } from "next/link";
+import { Post, Series } from "@/.contentlayer/generated";
 
 export interface PostHeading {
   heading: number;
@@ -35,4 +36,37 @@ export type SiteMetaData = {
   metadataBase: URL;
   newsletterUrl: string;
   analyticsProvider: AnalyticsProvider;
+};
+
+export type SeriesItem = {
+  title: string;
+  slug: Post["slug"];
+  status: Post["status"];
+  isCurrent: boolean;
+};
+
+export type PostSeries = Series & { posts: SeriesItem[] };
+
+export type PostWithSeries = Omit<Post, "series"> & { series: PostSeries };
+
+export type AuthorType = {
+  name: string;
+  handle: string;
+  social: {
+    github: string;
+    instagram: string;
+    linkedin: string;
+    tiktok: string;
+    twitter: string;
+    youtube: string;
+  };
+  email: string;
+  website: string;
+  jobTitle: string;
+  company: string;
+  availableForWork: boolean;
+  location: {
+    city: string;
+    media: string;
+  };
 };